Expand searched origin shifts for alternative settings
Candidates for matching alternative settings in MSG search are confined
to Affine normalizer of XSG or FSG. For trigonal and hexagonal space
groups, translations with 1/3 or 2/3 may be also allowed in addition to
the previously considered origin shifts. This commit expands the
searched origin shifts with denominator 3, but the results is the same
with the previous search.